refactor: rename new implementation (drop .zod) pt 2
