+++ lib/builder/clip_base.ts
@@ -1,8 +1,8 @@
 import { parse_unit } from '../parsers/unit.ts'
-import type * as parsed from '../parsers/template.zod.ts'
-import type { TimelineClip } from '../timeline.zod.ts'
-import type { ComputedGeometry } from '../geometry.zod.ts'
-import type { ClipInfo } from '../probe.zod.ts'
+import type * as parsed from '../parsers/template.ts'
+import type { TimelineClip } from '../timeline.ts'
+import type { ComputedGeometry } from '../geometry.ts'
+import type { ClipInfo } from '../probe.ts'
 
 import { type ClipBuilderData } from './ffmpeg_base.ts'
