+++ lib/builder/ffmpeg_sample.ts
@@ -1,10 +1,10 @@
 import * as path from '@std/path'
-import { parse_duration } from '../parsers/duration.zod.ts'
+import { parse_duration } from '../parsers/duration.ts'
 import { Context } from '../context.ts'
 import { relative_path } from '../util.ts'
-import type * as inputs from '../template_input.zod.ts'
-import type * as parsed from '../parsers/template.zod.ts'
-import type { ClipInfo } from '../probe.zod.ts'
+import type * as inputs from '../template_input.ts'
+import type * as parsed from '../parsers/template.ts'
+import type { ClipInfo } from '../probe.ts'
 
 import { FfmpegBuilderBase } from './ffmpeg_base.ts'
 import { ClipBuilderBase } from './clip_base.ts'
