+++ lib/cli.ts
@@ -7,11 +7,11 @@ import {open} from '@opensrc/deno-open'
 // import { open } from 'https://deno.land/x/open@v0.0.2/index.ts'
 // // import open from 'npm:open'
 import * as errors from './errors.ts'
-import { parse_template } from './parsers/template.zod.ts'
-import { render_video, render_sample_frame } from './mod.zod.ts'
+import { parse_template } from './parsers/template.ts'
+import { render_video, render_sample_frame } from './mod.ts'
 import { Logger } from './logger.ts'
 import { InstanceContext } from './context.ts'
-import type * as inputs from './template_input.zod.ts'
+import type * as inputs from './template_input.ts'
 import type { ContextOptions } from './context.ts'
