+++ lib/mod.ts
@@ -1,22 +1,22 @@
 import * as path from '@std/path'
 import * as errors from './errors.ts'
 import { parse_unit } from './parsers/unit.ts'
-import { parse_template } from './parsers/template.zod.ts'
+import { parse_template } from './parsers/template.ts'
 import { parse_percentage } from './parsers/unit.ts'
-import { parse_duration, fmt_human_readable_duration } from './parsers/duration.zod.ts'
+import { parse_duration, fmt_human_readable_duration } from './parsers/duration.ts'
 import { InstanceContext, Context } from './context.ts'
-import { compute_geometry, compute_size, compute_rotated_size } from './geometry.zod.ts'
-import { compute_zoompans } from './zoompan.zod.ts'
-import { compute_timeline } from './timeline.zod.ts'
-import { create_text_image } from './canvas/font.zod.ts'
+import { compute_geometry, compute_size, compute_rotated_size } from './geometry.ts'
+import { compute_zoompans } from './zoompan.ts'
+import { compute_timeline } from './timeline.ts'
+import { create_text_image } from './canvas/font.ts'
 import { relative_path } from './util.ts'
-import type * as inputs from './template_input.zod.ts'
-import type * as parsed from './parsers/template.zod.ts'
-import type { TimelineClip } from './timeline.zod.ts'
-import type { ComputedGeometry } from './geometry.zod.ts'
-import type { ClipInfo } from './probe.zod.ts'
+import type * as inputs from './template_input.ts'
+import type * as parsed from './parsers/template.ts'
+import type { TimelineClip } from './timeline.ts'
+import type { ComputedGeometry } from './geometry.ts'
+import type { ClipInfo } from './probe.ts'
 import type { ContextOptions } from './context.ts'
-import { ffmpeg } from './bindings/ffmpeg.zod.ts'
+import { ffmpeg } from './bindings/ffmpeg.ts'
 
 import { FfmpegBuilderBase, type ClipBuilderData } from './builder/ffmpeg_base.ts'
 import { ClipBuilderBase } from './builder/clip_base.ts'
@@ -150,5 +150,5 @@ async function render_image(template: inputs.Template, options: ContextOptions, 
 
 export { render_video, render_sample_frame, render_image }
 export type Template = inputs.Template
-export type { TemplateParsed } from './parsers/template.zod.ts'
+export type { TemplateParsed } from './parsers/template.ts'
 export type { RenderData } from './builder/ffmpeg_base.ts'
