+++ lib/probe.ts
@@ -3,11 +3,11 @@ import {readlines} from './util.ts'
 import { ProbeError, CommandError } from './errors.ts'
 import { AbstractClipMap } from './util.ts'
 import { parse_aspect_ratio, parse_ffmpeg_packet } from './parsers/ffmpeg_output.ts'
-import { compute_rotated_size } from './geometry.zod.ts'
+import { compute_rotated_size } from './geometry.ts'
 import type { InstanceContext } from './context.ts'
-import type * as template from './template_input.zod.ts'
-import type { MediaClipParsed } from './parsers/template.zod.ts'
-import type { Seconds } from './parsers/duration.zod.ts'
+import type * as template from './template_input.ts'
+import type { MediaClipParsed } from './parsers/template.ts'
+import type { Seconds } from './parsers/duration.ts'
 
 const CLIP_INFO_FILENAME = 'probe_info.json'
