+++ test/mod.test.ts
@@ -1,5 +1,5 @@
-import ffmpeg_templates from '../lib/cli.zod.ts'
-import { render_image, render_video } from '../lib/mod.zod.ts'
+import ffmpeg_templates from '../lib/cli.ts'
+import { render_image, render_video } from '../lib/mod.ts'
 import * as path from '@std/path'
 import { test } from './tools/test.ts'
